PDFBOX-4671 Replaced JUL service logger by SLF4J
Avoid problems with multiple web applications running on servlet
container.

Using a service provider for the Logger could arise NoClassDefFoundError
errors when an application it's stooped and started from the servlet
container manager.

Now, we simply use SLF4J as log API. Why? It's the facto standard
logger facade. Also, only it's necessary to replace the imports to use it on
the actual code.